[AC-5421] Initial refactor of Allocator to allow for multiple heuristics
