+++ allocate.py
@@ -9,31 +9,23 @@ import csv
 import sys
 from random import choice
 
-from classes.allocator import Allocator
-from classes.bin import BIN_DEFAULT_WEIGHT
-from classes.female_bin import FemaleBin
-from classes.home_program_bin import HomeProgramBin
-from classes.industry_bin import IndustryBin
-from classes.judge import Judge
-from classes.property import program
-from classes.property import industry
-from classes.reads_bin import ReadsBin
-from classes.role_bin import RoleBin
-from classes.satisfied_bin import SatisfiedBin
-from classes.startup import Startup
+from classes.allocator import (
+    Allocator,
+    RANDOM_SELECTION_HEURISTIC,
+)
 
+
+filename = None
+heuristic = RANDOM_SELECTION_HEURISTIC
 
 if len(sys.argv) > 1:
     filepath = sys.argv[1]
-else:
-    filepath = None
+if len(sys.argv) > 2:
+    heuristic = sys.argv[2]
 
 
-allocator = Allocator(filepath)
+allocator = Allocator(filepath, heuristic)
 allocator.read_entities()
-allocator.default_bins()
-allocator.add_startups()
-allocator.calc_capacity()
+allocator.setup()
 allocator.allocate()
-allocator.assess_allocations()
-allocator.assess_bins()
+allocator.assess()
