+++ classes/heuristic.py
@@ -1,0 +1,5 @@
+from classes.feature_bins import FeatureBins
+
+
+def find_heuristic(name):
+    return FeatureBins()
